refactored the notify bean
