+++ pkg/notifier/SESNotificationService.go
@@ -18,6 +18,7 @@ package notifier
 
 import (
 	"fmt"
+	"github.com/devtron-labs/devtron/client/events/bean"
 	"github.com/devtron-labs/devtron/internal/sql/repository"
 	"github.com/devtron-labs/devtron/internal/util"
 	"github.com/devtron-labs/devtron/pkg/notifier/adapter"
@@ -165,7 +166,7 @@ func (impl *SESNotificationServiceImpl) DeleteNotificationConfig(deleteReq *bean
 		impl.logger.Errorw("No matching entry found for delete", "err", err, "id", deleteReq.Id)
 		return err
 	}
-	notifications, err := impl.notificationSettingsRepository.FindNotificationSettingsByConfigIdAndConfigType(deleteReq.Id, beans.SES_CONFIG_TYPE)
+	notifications, err := impl.notificationSettingsRepository.FindNotificationSettingsByConfigIdAndConfigType(deleteReq.Id, bean.SES_CONFIG_TYPE)
 	if err != nil && err != pg.ErrNoRows {
 		impl.logger.Errorw("error in deleting ses config", "config", deleteReq)
 		return err
